requirements.txt
Flask>=3.0.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
Werkzeug>=2.3.0
